testing resource creating
