+++ packages/main-ui/src/pages/resources/components/drop_zone/mod.rs
@@ -3,7 +3,7 @@
 use dioxus::prelude::*;
 use dioxus_logger::tracing;
 use dioxus_translate::{translate, Language};
-use models::MetadataRequest;
+use models::{File, MetadataRequest};
 
 use crate::{
     components::{icons::UploadFile, upload_button::UploadButton},
@@ -18,8 +18,6 @@ use std::sync::Arc;
 
 mod i18n;
 use i18n::DropZoneTranslate;
-
-use super::create_resource_modal::File;
 
 fn human_readable_size(bytes: usize) -> String {
     let sizes = ["B", "KB", "MB", "GB", "TB"];
@@ -45,22 +43,25 @@ pub async fn handle_file_upload(file_engine: Arc<dyn FileEngine>) -> Vec<File> {
             Some(bytes) => {
                 let file_name: String = f.into();
                 let ext = file_name.rsplitn(2, '.').next().unwrap_or("");
-                let extension = ext.parse::<super::create_resource_modal::FileExtension>();
+                let extension = models::FileExtension::from_str(ext);
                 match extension {
                     Ok(ext) => {
-                        let url = api
+                        let url = match api
                             .upload_metadata(MetadataRequest {
                                 file_name: file_name.clone(),
                                 bytes: bytes.clone(),
                             })
-                            .await;
+                            .await
+                        {
+                            Ok(v) => Some(v),
+                            Err(_) => None,
+                        };
 
                         result.push(File {
                             name: file_name,
                             size: human_readable_size(bytes.len()),
-                            bytes,
                             ext,
-                            url: Some(url),
+                            url,
                         });
                     }
                     Err(_) => {
