+++ packages/main-ui/src/pages/resources/controller.rs
@@ -5,14 +5,14 @@ use dioxus::prelude::*;
 use dioxus_logger::tracing;
 use dioxus_translate::Language;
 use models::{
-    AccessLevel, ProjectArea, QueryResponse, ResourceCreateRequest, ResourceGetResponse,
+    AccessLevel, File, ProjectArea, QueryResponse, ResourceCreateRequest, ResourceGetResponse,
     ResourceQuery, ResourceSummary, ResourceType, ResourceUpdateRequest, Source, UsagePurpose,
 };
 
 use crate::{
     api, config,
     pages::resources::components::create_resource_modal::{
-        CreateResourceModal, File, ModifyResourceModal, RemoveResourceModal,
+        CreateResourceModal, ModifyResourceModal, RemoveResourceModal,
     },
     service::{
         login_service::LoginService,
@@ -187,6 +187,7 @@ impl Controller {
         usage_purpose: Option<UsagePurpose>,
         source: Option<Source>,
         access_level: Option<AccessLevel>,
+        files: Vec<File>,
     ) -> Result<(), models::ApiError> {
         let org = self.user.get_selected_org();
         if org.is_none() {
@@ -203,6 +204,7 @@ impl Controller {
                 usage_purpose,
                 source,
                 access_level,
+                files,
             )
             .await
         {
@@ -231,6 +233,7 @@ impl Controller {
                 resource.usage_purpose,
                 resource.source,
                 resource.access_level,
+                resource.files,
             )
             .await
         {
@@ -258,7 +261,7 @@ impl Controller {
             .open(rsx! {
                 CreateResourceModal {
                     lang,
-                    onupload: move |(title, resource_type, field, purpose, source, access_level, _files)| {
+                    onupload: move |(title, resource_type, field, purpose, source, access_level, files)| {
                         async move {
                             ctrl.create_resource(
                                     title,
@@ -267,6 +270,7 @@ impl Controller {
                                     purpose,
                                     source,
                                     access_level,
+                                    files,
                                 )
                                 .await;
                         }
